+++ GenericModConfigMenu/Framework/UI/Table.cs
@@ -1,6 +1,8 @@
+using System;
 using System.Collections.Generic;
 using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
+using StardewValley;
 using StardewValley.Menus;
 
 namespace GenericModConfigMenu.Framework.UI
@@ -114,21 +116,33 @@ namespace GenericModConfigMenu.Framework.UI
         /// <inheritdoc />
         public override void Draw(SpriteBatch b)
         {
-            IClickableMenu.drawTextureBox(b, (int)this.Position.X - 32, (int)this.Position.Y - 32, (int)this.Size.X + 64, (int)this.Size.Y + 64, Color.White);
+            // calculate draw area
+            var backgroundArea = new Rectangle((int)this.Position.X - 32, (int)this.Position.Y - 32, (int)this.Size.X + 64, (int)this.Size.Y + 64);
+            int contentPadding = 12;
+            var contentArea = new Rectangle(backgroundArea.X + contentPadding, backgroundArea.Y + contentPadding, backgroundArea.Width - contentPadding * 2, backgroundArea.Height - contentPadding * 2);
 
-            foreach (var row in this.Rows)
+            // draw background
+            IClickableMenu.drawTextureBox(b, backgroundArea.X, backgroundArea.Y, backgroundArea.Width, backgroundArea.Height, Color.White);
+
+            // draw table contents
+            // This uses a scissor rectangle to clip content taller than one row that might be
+            // drawn past the bottom of the UI, like images or complex options.
+            this.InScissorRectangle(b, contentArea, contentBatch =>
             {
-                foreach (var element in row)
+                foreach (var row in this.Rows)
                 {
-                    if (element.Position.Y < this.Position.Y || element.Position.Y + this.RowHeight - Table.RowPadding > this.Position.Y + this.Size.Y)
-                        continue;
-                    if (element == this.RenderLast)
-                        continue;
-                    element.Draw(b);
+                    foreach (var element in row)
+                    {
+                        if (element.Position.Y < this.Position.Y || element.Position.Y + this.RowHeight - Table.RowPadding > this.Position.Y + this.Size.Y)
+                            continue;
+                        if (element == this.RenderLast)
+                            continue;
+                        element.Draw(contentBatch);
+                    }
                 }
-            }
 
-            this.RenderLast?.Draw(b);
+                this.RenderLast?.Draw(contentBatch);
+            });
 
             this.Scrollbar.Draw(b);
         }
@@ -143,6 +157,35 @@ namespace GenericModConfigMenu.Framework.UI
             this.Scrollbar.RequestHeight = (int)this.Size.Y;
             this.Scrollbar.Rows = this.Rows.Count;
             this.Scrollbar.FrameSize = (int)(this.Size.Y / this.RowHeight);
+        }
+
+        private void InScissorRectangle(SpriteBatch spriteBatch, Rectangle area, Action<SpriteBatch> draw)
+        {
+            // render the current sprite batch to the screen
+            spriteBatch.End();
+
+            // start temporary sprite batch
+            using SpriteBatch contentBatch = new SpriteBatch(Game1.graphics.GraphicsDevice);
+            GraphicsDevice device = Game1.graphics.GraphicsDevice;
+            Rectangle prevScissorRectangle = device.ScissorRectangle;
+
+            // render in scissor rectangle
+            try
+            {
+                device.ScissorRectangle = area;
+                contentBatch.Begin(SpriteSortMode.Deferred, BlendState.NonPremultiplied, SamplerState.PointClamp, null, Utility.ScissorEnabled);
+
+                draw(contentBatch);
+
+                contentBatch.End();
+            }
+            finally
+            {
+                device.ScissorRectangle = prevScissorRectangle;
+            }
+
+            // resume previous sprite batch
+            spriteBatch.Begin();
         }
     }
 }